Match argument types between python and C++

+++ src/linear_regression.cpp
@@ -10,7 +10,7 @@
 const std::string VERSION = "1.0.0";
 
 // Wrapper around fitKmer block which multithreads calculation
-void fitKmers(const py::array_t<double, py::array::c_style | py::array::forcecast>& raw,
+void fitKmers(const py::array_t<float, py::array::c_style | py::array::forcecast>& raw,
               py::array_t<double, py::array::c_style | py::array::forcecast>& dists,
               const column_vector& klist,
               const int num_threads)
@@ -46,7 +46,7 @@ void fitKmers(const py::array_t<double, py::array::c_style | py::array::forcecas
     }
 }
 
-void fitKmerBlock(const py::array_t<double, py::array::c_style | py::array::forcecast>& raw,
+void fitKmerBlock(const py::array_t<float, py::array::c_style | py::array::forcecast>& raw,
                   py::array_t<double, py::array::c_style | py::array::forcecast>& dists,
                   const column_vector& klist,
                   const size_t start,
